fix error when shutting down receiving node
shutdown(socket, SHUT_RD) causes recvfrom to return with value 0 leading the
program to falsely assume a packet with 0 bytes was received and call the
recv_handle leading to a deserialization error. checking for packet size and
only calling recv_handle when bytes have been received allows the program to
avoid this issue and gracefully exit.