add message relay prioritization
As an approximation of true robust ACKs, nodes delay relaying messages according
to their priority (order) in the header relays list. Messages to be delayed are
requeued and processed later, allowing time for higher priority relays to
transmit. The queued relay is then cancelled if the transmission from a higher
priority relay is intercepted.